add ru calculator integration test for torch
